Add the login page
- Use Auth.js for authentication and authorization.
- Move all actions to one file.
- Remove unused static resources.

+++ lib/actions.ts
@@ -3,6 +3,7 @@
 import prisma from '@/lib/db'
 import { revalidatePath } from 'next/cache'
 import { z } from 'zod'
+import { auth, signIn } from './auth'
 
 const createWorkSchema = z.object({
   name: z.string().min(1),
@@ -13,6 +14,11 @@ const createWorkSchema = z.object({
 })
 
 export async function createWork(formData: FormData) {
+  const session = await auth()
+  if (!session?.user) {
+    throw Error('You must be signed in to perform this action.')
+  }
+
   const validatedFields = createWorkSchema.safeParse(
     Object.fromEntries(formData)
   )
@@ -24,6 +30,23 @@ export async function createWork(formData: FormData) {
 }
 
 export async function deleteWork(id: string) {
+  const session = await auth()
+  if (!session?.user) {
+    throw Error('You must be signed in to perform this action.')
+  }
+
   await prisma.work.deleteMany({ where: { id } })
   revalidatePath('/dashboard')
+}
+
+const loginSchema = z.object({
+  provider: z.enum(['github']),
+})
+
+export async function login(formData: FormData) {
+  const validatedFields = loginSchema.safeParse(Object.fromEntries(formData))
+  if (!validatedFields.success) {
+    return { errors: validatedFields.error.flatten().fieldErrors }
+  }
+  await signIn(validatedFields.data.provider)
 }
